#6519 – Proper split of tabs/accordion
