+++ packages/ketcher-macromolecules/src/components/monomerLibrary/RnaBuilder/RnaElementsView/Summary/Summary.tsx
