+++ packages/ketcher-macromolecules/src/components/monomerLibrary/RnaBuilder/RnaElementsView/Summary/types.ts
